+++ crates/compilers/src/preprocessor.rs
@@ -88,9 +88,13 @@ pub(crate) fn interface_representation(
         return Err(err);
     }
 
-    let content =
-        replace_source_content(content, spans_to_remove.iter().map(|span| (span.to_range(), "")))
-            .replace("\n", "");
+    let content = replace_source_content(
+        content,
+        spans_to_remove
+            .iter()
+            .map(|span| (span.to_range().start, span.to_range().end, String::new())),
+    )
+    .replace("\n", "");
     Ok(utils::RE_TWO_OR_MORE_SPACES.replace_all(&content, "").to_string())
 }
